+++ sketch-core/src/main/kotlin/com/github/panpf/sketch/util/Utils.kt
@@ -48,7 +48,7 @@ import java.text.DecimalFormat
 import kotlin.coroutines.resume
 
 
-inline fun <R> ifOrNull(value: Boolean, block: () -> R?): R? = if (value) block() else null
+internal inline fun <R> ifOrNull(value: Boolean, block: () -> R?): R? = if (value) block() else null
 
 /**
  * Convert to the type specified by the generic, if this is null or cannot be converted return null