+++ sketch-extensions-core/src/androidTest/kotlin/com/github/panpf/sketch/extensions/core/test/drawable/AbsProgressDrawableTest.kt
@@ -1,0 +1,9 @@
+package com.github.panpf.sketch.extensions.core.test.drawable
+
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import org.junit.runner.RunWith
+
+@RunWith(AndroidJUnit4::class)
+class AbsProgressDrawableTest {
+    // todo Waiting for test
+}