+++ sketch-extensions-core/src/main/kotlin/com/github/panpf/sketch/drawable/SectorProgressDrawable.kt
@@ -38,7 +38,11 @@ class SectorProgressDrawable(
     hiddenWhenIndeterminate: Boolean = false,
     hiddenWhenCompleted: Boolean = true,
     stepAnimationDuration: Int = DEFAULT_STEP_ANIMATION_DURATION,
-) : AbsProgressDrawable(hiddenWhenIndeterminate, hiddenWhenCompleted, stepAnimationDuration) {
+) : AbsProgressDrawable(
+    hiddenWhenIndeterminate = hiddenWhenIndeterminate,
+    hiddenWhenCompleted = hiddenWhenCompleted,
+    stepAnimationDuration = stepAnimationDuration
+) {
 
     companion object {
         fun defaultSize(): Int {